requirements.txt
Django==3.2
pillow==9.2.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.51
djangorestframework==3.12.4
psycopg2-binary==2.9.3